SR-5892. [Ситиматик] При загрузке ПЛ в мобильном приложение ошибка - "Проблема с интернетом" https://jira.smartro.ru/browse/SR-5892
1 of 1.

+++ app/src/main/java/ru/smartro/worknote/work/WorkOrderResponse_know1.kt
@@ -96,7 +96,7 @@ data class Platform_know1(
     @SerializedName("updated_at")
     var updateAt: Long,
     @SerializedName("srp_id")
-    val srpId: Int,
+    val srpId: Long,
     @SerializedName("order_start_time")
     var orderStartTime: String? = null,
     @SerializedName("order_end_time")
